requirements.txt
aiosqlite==0.21.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.3
asgi-lifespan==2.1.0
asgiref==3.8.1
asyncpg==0.30.0
attrs==25.3.0
beautifulsoup4==4.13.3
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coolname==2.2.0
cryptography==44.0.2
dateparser==1.2.1
Deprecated==1.2.18
dj-database-url==2.2.0
Django==5.1
django-appconf==1.0.6
django-browser-reload==1.15.0
django-vite==3.0.4
djangorestframework==3.15.2
docker==7.1.0
environs==11.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
fsspec==2025.3.2
google-api-core==2.24.1
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-bigquery==3.29.0
google-cloud-bigquery-storage==2.30.0
google-cloud-core==2.4.3
google-cloud-secret-manager==2.23.2
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.67.0
graphviz==0.20.3
greenlet==3.1.1
griffe==1.7.2
grpc-google-iam-v1==0.14.2
grpcio==1.70.0
grpcio-status==1.70.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.2
hyperframe==6.1.0
hypothesis==6.131.16
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==3.0.2
marshmallow==3.22.0
mdurl==0.1.2
numpy==2.1.1
oauthlib==3.2.2
opentelemetry-api==1.31.1
orjson==3.10.16
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pendulum==3.1.0
pluggy==1.5.0
prefect==3.3.3
prefect-gcp==0.6.5
prometheus_client==0.21.1
proto-plus==1.26.0
protobuf==5.29.3
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.11.2
pydantic-extra-types==2.10.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.3.4
pytest-django==4.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.1
PyYAML==6.0.2
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.24.0
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
schema==0.7.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.1
starlette==0.46.1
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uv==0.6.12
uvicorn==0.34.0
websockets==15.0.1
whenever==0.7.3
whitenoise==6.6.0
wrapt==1.17.2
zipp==3.21.0